+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Moved `ItemContext` to react folder.
+
 ## [0.19.1] - 2020-03-16
 
 ## [0.19.0] - 2020-03-09